Restructure the code.
