+++ src/model.rs
@@ -1,6 +1,3 @@
-mod borrowed;
-mod owned;
-
 use std::boxed::Box;
 use std::slice::Iter;
