--- src/model/borrowed/mod.rs
@@ -1,7 +1,0 @@
-mod v1;
-mod v2;
-
-pub enum Header<'a> {
-    VersionOne(v1::Header<'a>),
-    VersionTwo(v2::Header<'a>),
-}
