--- src/model/owned.rs
@@ -1,1 +1,0 @@
-
